+++ apps/omsorgspengesoknad/src/app/pages/velkommen/VelkommenPage.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryFn } from '@storybook/react-vite';
 import VelkommenPage from './VelkommenPage';
 import { withIntl } from '../../../storybook/decorators/withIntl';
 import { withRouterProvider } from '../../../storybook/decorators/withRouter';
-import { withAmplitudeProvider } from '../../../storybook/decorators/withAmplitudeProvider';
+import { withAnalyticsProvider } from '../../../storybook/decorators/withAnalyticsProvider';
 import { withSøknadContextProvider } from '../../../storybook/decorators/withSøknadContext';
 import { velkommenPageMessages } from './velkommenPageMessages';
 import MessagesPreview from '@navikt/sif-common-core-ds/src/dev-utils/intl/messages-preview/MessagesPreview';
@@ -11,7 +11,7 @@ import { Box, Tabs } from '@navikt/ds-react';
 export default {
     title: 'Pages/Velkommen',
     component: VelkommenPage,
-    decorators: [withIntl, withRouterProvider, withAmplitudeProvider, withSøknadContextProvider],
+    decorators: [withIntl, withRouterProvider, withAnalyticsProvider, withSøknadContextProvider],
 } as Meta<typeof VelkommenPage>;
 
 const Template: StoryFn<typeof VelkommenPage> = () => (