+++ apps/opplaringspenger-soknad/src/app/pages/kvittering/KvitteringPage.stories.tsx
@@ -1,13 +1,13 @@
 import { Meta, StoryFn } from '@storybook/react-vite';
 import KvitteringPage from '../../../app/pages/kvittering/KvitteringPage';
-import { withAmplitudeProvider } from '../../../storybook/decorators/withAmplitude';
+import { withAnalyticsProvider } from '../../../storybook/decorators/withAnalyticsProvider';
 import { withIntl } from '../../../storybook/decorators/withIntl';
 import { withRouterProvider } from '../../../storybook/decorators/withRouter';
 
 export default {
     title: 'Pages/KvitteringPage',
     component: KvitteringPage,
-    decorators: [withIntl, withRouterProvider, withAmplitudeProvider],
+    decorators: [withIntl, withRouterProvider, withAnalyticsProvider],
 } as Meta<typeof KvitteringPage>;
 
 const Template: StoryFn<typeof KvitteringPage> = () => <KvitteringPage />;