+++ packages/sif-app-register/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/sif-app-register
 
+## 0.1.12
+
+### Patch Changes
+
+- Erstatte begrepet amplitude med analytics. Ingen logikk forandring. Dette er på grunn av at amplitude fases ut.
+
 ## 0.1.11
 
 ### Patch Changes