+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @navikt/sif-common-forms-ds
 
+## 21.0.34
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-env@0.1.9
+    - @navikt/sif-common-ui@0.11.34
+    - @navikt/sif-common-core-ds@10.3.33
+
 ## 21.0.33
 
 ### Patch Changes