+++ packages/sif-common-soknad-ds/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @navikt/sif-common-soknad-ds
 
+## 26.1.16
+
+### Patch Changes
+
+- Erstatte begrepet amplitude med analytics. Ingen logikk forandring. Dette er på grunn av at amplitude fases ut.
+- Updated dependencies
+    - @navikt/sif-common-env@0.1.9
+    - @navikt/sif-common-ui@0.11.34
+    - @navikt/sif-common-core-ds@10.3.33
+    - @navikt/sif-common-forms-ds@21.0.34
+
 ## 26.1.15
 
 ### Patch Changes